Utworzono folder Forms na formularze. Dodano komentarze do FormActivity i zrobiono reformat.
