+++ ActivityMonitor/App.xaml
@@ -2,7 +2,7 @@
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:local="clr-namespace:ActivityMonitor"
-             StartupUri="FormActivity.xaml">
+             StartupUri="Forms/FormActivity.xaml">
     <Application.Resources>
          
     </Application.Resources>
